requirements.txt
asgiref==3.8.1
Django==3.2.4
djangorestframework==3.12.4
pytz==2025.1
sqlparse==0.5.3
